Separate 'Ipc' from 'WebSocket' tests
